+++ Remote/src/main/java/com/breadtask/remote/RemoteDataSource.kt
@@ -5,7 +5,7 @@ import javax.inject.Inject
 
 class RemoteDataSource @Inject constructor(val service: PostsService) : IRemoteDataSource {
     override suspend fun fetchPosts(page: Int?, perPage: Int?) = service.getPosts(page, perPage)
-    override suspend fun getCommentsForPost(postId: Int, page: Int?, perPage: Int?) =
+    override suspend fun getCommentsForPost(postId: Long, page: Int?, perPage: Int?) =
         service.getCommentsForPost(postId, page, perPage)
 
 }
